Add tap to click support with i3
Added a command to set the xinput property for my laptop's touchpad to
be able to click on tap.
Should create another file specifically for xinput, but it isn't read in
the home directory, so for now let's make do with i3, and I'll deal with
this when I'm not lazy.